+++ src/components/accountComponent/AccountSlider.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { Card, CardContent, Button, Box, Typography, IconButton } from "@mui/material";
 import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
 import ChevronRightIcon from "@mui/icons-material/ChevronRight";
-import { fetchAccountsId } from './services/axios';
+import {fetchAccountsId} from "../../services/AxiosBanking";
 
 const AccountSlider = () => {
     const [currentIndex, setCurrentIndex] = useState(0);
